(chore) fix coding standards
