+++ tests/TestCase/Controller/Component/ImagineComponentTest.php
@@ -11,65 +11,11 @@ declare(strict_types=1);
  */
 namespace Burzum\Imagine\Test\TestCase\Controller\Component;
 
-use Cake\Controller\Controller;
+use Burzum\Imagine\Test\App\Controller\ImagineImagesTestController;
 use Cake\Core\Configure;
-use Cake\Event\EventInterface;
 use Cake\Http\Response;
 use Cake\Http\ServerRequest as Request;
 use Cake\TestSuite\TestCase;
-
-/**
- * ImagineImagesTestController
- *
- * @property ImagineComponent $Imagine
- */
-class ImagineImagesTestController extends Controller
-{
-    /**
-     * @var string
-     */
-    public $name = 'Images';
-
-    /**
-     * @var array
-     */
-    public $uses = ['Images'];
-
-    /**
-     * Redirect url
-     * @var mixed
-     */
-    public $redirectUrl = null;
-
-    /**
-     * @inheritdoc
-     * @throws \Exception
-     */
-    public function initialize(): void
-    {
-        parent::initialize();
-        $this->loadComponent('Burzum/Imagine.Imagine');
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function beforeFilter(EventInterface $Event)
-    {
-        parent::beforeFilter($Event);
-        $this->Imagine->userModel = 'UserModel';
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function redirect($url, int $status = 302): ?Response
-    {
-        $this->redirectUrl = $url;
-
-        return null;
-    }
-}
 
 /**
  * Imagine Component Test
